Move types tests to src
