+++ packages/types/src/Hash256.test.ts
@@ -1,6 +1,6 @@
 import { expect } from 'earljs'
 
-import { Hash256 } from '../src'
+import { Hash256 } from './Hash256'
 
 describe(Hash256.name, () => {
   it('accepts lowercase hashes', () => {
